Update test case and main plugin routines
The test case is now such that when compiled with -O3 generates
all instructions of interest for the plugin. This is with gcc 8.2.1.

In this way, longer asm volatile blocks won't be supported by the
plugin but, in the end, this can make sense (isn't asm volatile
something which shouldn't be managed by the compiler, after all?)

Also, the RTL pass is now printing only RTL instructions which
might be of interest for our instrumentation purpose.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>